Add blog DTO, hydrator and update repository methods
The code adds a DTO (Data Transfer Object) for the `Blog` entity and a `BlogHydrator` class for hydrating data from the database into this DTO. It also updates `BlogRepository` to retrieve data as DTOs. Additionally, it includes Symfony PropertyAccess package for managing accessing properties of objects. The purpose of this setup is to improve performance by avoiding unnecessary hydration of Doctrine entities during API data rendering.